# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user profile retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get current user using JWT",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Current user updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update current user using JWT",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/me/preferences": {
      "put": {
        "operationId": "UsersController_updatePreferences",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preferences updated successfully."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update current user preferences using JWT",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/teams": {
      "post": {
        "operationId": "TeamsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Team created successfully"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create a new team",
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{teamId}/invite": {
      "post": {
        "operationId": "TeamsController_inviteMember",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Member invited successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Invite a member to the team",
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{teamId}/members/{memberId}": {
      "put": {
        "operationId": "TeamsController_updateMemberRole",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Member role updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update a team member’s role",
        "tags": [
          "Teams"
        ]
      },
      "delete": {
        "operationId": "TeamsController_removeMember",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Remove a member from the team",
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{teamId}/posts/{postId}/approve": {
      "post": {
        "operationId": "TeamsController_approvePost",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Approve a post for publishing",
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{teamId}/posts/{postId}/reject": {
      "post": {
        "operationId": "TeamsController_rejectPost",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Reject a post",
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{teamId}/audit-logs": {
      "get": {
        "operationId": "TeamsController_listAuditLogs",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamAuditLogResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List team audit logs",
        "tags": [
          "Teams"
        ]
      }
    },
...
    "/webhooks": {
      "post": {
        "operationId": "WebhooksController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Register webhook",
        "tags": [
          "Webhooks"
        ]
      },
      "get": {
        "operationId": "WebhooksController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List webhooks",
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/webhooks/{webhookId}": {
      "delete": {
        "operationId": "WebhooksController_remove",
        "parameters": [
          {
            "name": "webhookId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Delete webhook",
        "tags": [
          "Webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "YSHAI API",
    "description": "Generated for Orval client",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          },
          "timezone": {
            "type": "string"
          },
          "isEmailVerified": {
            "type": "boolean"
          },
          "language": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "timeFormat": {
            "type": "string",
            "enum": [
              "12h",
              "24h"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "isEmailVerified",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "timezone": {
            "type": "string"
          },
          "timeFormat": {
            "type": "string",
            "enum": [
              "12h",
              "24h"
            ]
          },
          "language": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          }
        }
      },
      "UpdatePreferencesDto": {
        "type": "object",
        "properties": {
          "timezone": {
            "type": "string",
            "description": "IANA timezone, e.g., America/New_York"
          },
          "language": {
            "type": "string",
            "description": "BCP 47 language tag, e.g., en or ar"
          },
          "locale": {
            "type": "string",
            "description": "Locale for formatting, e.g., en-US or ar-EG"
          },
          "timeFormat": {
            "type": "string",
            "description": "12h or 24h clock format"
          }
        }
      },
...
      "CreateTeamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the team",
            "example": "Marketing Team"
          }
        },
        "required": [
          "name"
        ]
      },
      "InviteMemberDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email of the user to invite",
            "example": "member@example.com"
          },
          "role": {
            "type": "string",
            "description": "Role to assign to the invited member",
            "enum": [
              "admin",
              "editor",
              "viewer"
            ],
            "example": "editor"
          }
        },
        "required": [
          "email",
          "role"
        ]
      },
      "MessageDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "UpdateMemberRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "description": "New role for the team member",
            "enum": [
              "admin",
              "editor",
              "viewer"
            ],
            "example": "admin"
          }
        },
        "required": [
          "role"
        ]
      },
      "TeamAuditLogResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "action": {
            "type": "string"
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "action",
          "timestamp"
        ]
      },
...
      "CreateWebhookDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "HTTPS endpoint to receive events"
          },
          "event": {
            "type": "string",
            "enum": [
              "post.published",
              "post.failed",
              "account.disconnected"
            ]
          },
          "secret": {
            "type": "string",
            "description": "Shared secret to sign webhook payloads"
          }
        },
        "required": [
          "url",
          "event",
          "secret"
        ]
      },
...
    }
  }
}